# backend/openapi.json
{
 "components": {
  "schemas": {
...
   "CoverageItem": {
    "properties": {
     "category": {
      "title": "Category",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "missing": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Missing"
     },
     "proposal": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Proposal"
     },
     "sources": {
      "items": {
       "$ref": "#/components/schemas/CoverageSource"
      },
      "title": "Sources",
      "type": "array"
     },
     "status": {
      "enum": [
       "available",
       "partial",
       "log_only",
       "missing"
      ],
      "title": "Status",
      "type": "string"
     },
     "summary": {
      "title": "Summary",
      "type": "string"
     }
    },
    "required": [
     "id",
     "category",
     "status",
     "summary",
     "sources",
     "missing",
     "proposal"
    ],
    "title": "CoverageItem",
    "type": "object"
   },
   "CoverageSource": {
    "properties": {
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     },
     "found": {
      "title": "Found",
      "type": "boolean"
     },
     "kind": {
      "enum": [
       "tracepoint",
       "debugfs",
       "nl80211",
       "log",
       "sysfs",
       "none"
      ],
      "title": "Kind",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "role": {
      "enum": [
       "primary",
       "secondary",
       "log"
      ],
      "title": "Role",
      "type": "string"
     },
     "tier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier"
     }
    },
    "required": [
     "kind",
     "name",
     "role",
     "found",
     "detail",
     "tier"
    ],
    "title": "CoverageSource",
    "type": "object"
   },
   "DebugfsEntry": {
    "properties": {
     "path": {
      "title": "Path",
      "type": "string"
     },
     "policy": {
      "title": "Policy",
      "type": "string"
     },
     "readable": {
      "title": "Readable",
      "type": "boolean"
     },
     "tier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier"
     }
    },
    "required": [
     "path",
     "tier",
     "policy",
     "readable"
    ],
    "title": "DebugfsEntry",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Instrumentation": {
    "properties": {
     "custom_instrumentation": {
      "items": {
       "type": "string"
      },
      "title": "Custom Instrumentation",
      "type": "array"
     },
     "debugfs": {
      "items": {
       "$ref": "#/components/schemas/DebugfsEntry"
      },
      "title": "Debugfs",
      "type": "array"
     },
     "debugfs_truncated": {
      "default": false,
      "title": "Debugfs Truncated",
      "type": "boolean"
     },
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "driver": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Driver"
     },
     "generated_at": {
      "title": "Generated At",
      "type": "number"
     },
     "helper_available": {
      "title": "Helper Available",
      "type": "boolean"
     },
     "helper_note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Helper Note"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/CoverageItem"
      },
      "title": "Items",
      "type": "array"
     },
     "kernel_release": {
      "title": "Kernel Release",
      "type": "string"
     },
     "patches": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Patches",
      "type": "array"
     },
     "phy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phy"
     },
     "tracepoints": {
      "items": {
       "$ref": "#/components/schemas/TracepointInfo"
      },
      "title": "Tracepoints",
      "type": "array"
     }
    },
    "required": [
     "device_id",
     "phy",
     "driver",
     "kernel_release",
     "helper_available",
     "helper_note",
     "tracepoints",
     "debugfs",
     "debugfs_truncated",
     "items",
     "patches",
     "custom_instrumentation",
     "generated_at"
    ],
    "title": "Instrumentation",
    "type": "object"
   },
...
   "LinkActivity": {
    "properties": {
     "last_tx_rate_hex": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Tx Rate Hex"
     },
     "link": {
      "title": "Link",
      "type": "integer"
     },
     "rx_decrypted": {
      "default": 0,
      "title": "Rx Decrypted",
      "type": "integer"
     },
     "rx_dropped_flags": {
      "default": 0,
      "title": "Rx Dropped Flags",
      "type": "integer"
     },
     "rx_frames": {
      "default": 0,
      "title": "Rx Frames",
      "type": "integer"
     },
     "tx_ack_errors": {
      "default": 0,
      "title": "Tx Ack Errors",
      "type": "integer"
     },
     "tx_acked": {
      "default": 0,
      "title": "Tx Acked",
      "type": "integer"
     },
     "tx_completions": {
      "default": 0,
      "title": "Tx Completions",
      "type": "integer"
     },
     "tx_selected": {
      "default": 0,
      "title": "Tx Selected",
      "type": "integer"
     },
     "txfree_attempts_total": {
      "default": 0,
      "title": "Txfree Attempts Total",
      "type": "integer"
     },
     "txfree_failed": {
      "default": 0,
      "title": "Txfree Failed",
      "type": "integer"
     }
    },
    "required": [
     "link",
     "tx_selected",
     "tx_completions",
     "tx_acked",
     "tx_ack_errors",
     "txfree_failed",
     "txfree_attempts_total",
     "rx_frames",
     "rx_decrypted",
     "rx_dropped_flags",
     "last_tx_rate_hex"
    ],
    "title": "LinkActivity",
    "type": "object"
   },
...
   "Migration": {
    "properties": {
     "flow": {
      "title": "Flow",
      "type": "string"
     },
     "from_link": {
      "title": "From Link",
      "type": "integer"
     },
     "orig_link": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Orig Link"
     },
     "tid": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tid"
     },
     "to_link": {
      "title": "To Link",
      "type": "integer"
     },
     "ts_boottime_ns": {
      "title": "Ts Boottime Ns",
      "type": "integer"
     }
    },
    "required": [
     "ts_boottime_ns",
     "flow",
     "tid",
     "from_link",
     "to_link",
     "orig_link"
    ],
    "title": "Migration",
    "type": "object"
   },
...
   "Mt76Snapshot": {
    "properties": {
     "boottime_ns": {
      "title": "Boottime Ns",
      "type": "integer"
     },
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "include_wake": {
      "title": "Include Wake",
      "type": "boolean"
     },
     "link_stats": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Link Stats",
      "type": "array"
     },
     "mlo_active_link_ids": {
      "items": {
       "type": "integer"
      },
      "title": "Mlo Active Link Ids",
      "type": "array"
     },
     "mlo_active_links": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mlo Active Links"
     },
     "mlo_str_cap": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mlo Str Cap"
     },
     "phy": {
      "title": "Phy",
      "type": "string"
     },
     "reads": {
      "items": {
       "$ref": "#/components/schemas/ReadRecord"
      },
      "title": "Reads",
      "type": "array"
     },
     "runtime_pm": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Runtime Pm",
      "type": "object"
     },
     "rx_queues": {
      "items": {
       "additionalProperties": {
        "type": "integer"
       },
       "type": "object"
      },
      "title": "Rx Queues",
      "type": "array"
     },
     "settings": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Settings",
      "type": "object"
     },
     "skipped": {
      "items": {
       "additionalProperties": {
        "type": "string"
       },
       "type": "object"
      },
      "title": "Skipped",
      "type": "array"
     },
     "taken_at": {
      "title": "Taken At",
      "type": "number"
     },
     "tx_stats": {
      "additionalProperties": true,
      "title": "Tx Stats",
      "type": "object"
     },
     "wcid_dump": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Wcid Dump",
      "type": "array"
     },
     "xmit_queues": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Xmit Queues",
      "type": "array"
     }
    },
    "required": [
     "device_id",
     "phy",
     "taken_at",
     "boottime_ns",
     "include_wake",
     "reads",
     "xmit_queues",
     "rx_queues",
     "runtime_pm",
     "mlo_active_links",
     "mlo_active_link_ids",
     "mlo_str_cap",
     "link_stats",
     "wcid_dump",
     "tx_stats",
     "settings",
     "skipped"
    ],
    "title": "Mt76Snapshot",
    "type": "object"
   },
...
   "ReadRecord": {
    "properties": {
     "bytes": {
      "default": 0,
      "title": "Bytes",
      "type": "integer"
     },
     "elapsed_ms": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Elapsed Ms"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "ok": {
      "title": "Ok",
      "type": "boolean"
     },
     "path": {
      "title": "Path",
      "type": "string"
     },
     "raw": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Raw"
     },
     "read_at_boottime_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Read At Boottime Ns"
     },
     "tier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tier"
     }
    },
    "required": [
     "path",
     "tier",
     "ok",
     "error",
     "bytes",
     "elapsed_ms",
     "read_at_boottime_ns",
     "raw"
    ],
    "title": "ReadRecord",
    "type": "object"
   },
...
   "TagSummary": {
    "properties": {
     "attributed": {
      "default": 0,
      "title": "Attributed",
      "type": "integer"
     },
     "count": {
      "title": "Count",
      "type": "integer"
     },
     "last_fields": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Last Fields",
      "type": "object"
     },
     "last_summary": {
      "title": "Last Summary",
      "type": "string"
     },
     "last_ts_boottime_ns": {
      "title": "Last Ts Boottime Ns",
      "type": "integer"
     },
     "tag": {
      "title": "Tag",
      "type": "string"
     },
     "unattributed": {
      "default": 0,
      "title": "Unattributed",
      "type": "integer"
     }
    },
    "required": [
     "tag",
     "count",
     "attributed",
     "unattributed",
     "last_ts_boottime_ns",
     "last_fields",
     "last_summary"
    ],
    "title": "TagSummary",
    "type": "object"
   },
...
   "TidLinkCell": {
    "properties": {
     "aggr_count": {
      "default": 0,
      "title": "Aggr Count",
      "type": "integer"
     },
     "count": {
      "title": "Count",
      "type": "integer"
     },
     "link": {
      "title": "Link",
      "type": "integer"
     },
     "tid": {
      "title": "Tid",
      "type": "integer"
     }
    },
    "required": [
     "tid",
     "link",
     "count",
     "aggr_count"
    ],
    "title": "TidLinkCell",
    "type": "object"
   },
...
   "TraceAggregate": {
    "properties": {
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "events_considered": {
      "title": "Events Considered",
      "type": "integer"
     },
     "first_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "First Ns"
     },
     "last_ns": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Ns"
     },
     "links": {
      "items": {
       "$ref": "#/components/schemas/LinkActivity"
      },
      "title": "Links",
      "type": "array"
     },
     "migrations": {
      "items": {
       "$ref": "#/components/schemas/Migration"
      },
      "title": "Migrations",
      "type": "array"
     },
     "notes": {
      "items": {
       "type": "string"
      },
      "title": "Notes",
      "type": "array"
     },
     "path_counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Path Counts",
      "type": "object"
     },
     "tid_link": {
      "items": {
       "$ref": "#/components/schemas/TidLinkCell"
      },
      "title": "Tid Link",
      "type": "array"
     }
    },
    "required": [
     "device_id",
     "events_considered",
     "first_ns",
     "last_ns",
     "tid_link",
     "links",
     "migrations",
     "path_counts",
     "notes"
    ],
    "title": "TraceAggregate",
    "type": "object"
   },
...
   "TracepointInfo": {
    "properties": {
     "fields": {
      "items": {
       "type": "string"
      },
      "title": "Fields",
      "type": "array"
     },
     "group": {
      "title": "Group",
      "type": "string"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "print_fmt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Print Fmt"
     }
    },
    "required": [
     "group",
     "name",
     "fields",
     "print_fmt"
    ],
    "title": "TracepointInfo",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Wireless Hardware Debugger",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/devices/{device_id}/mt76/driver-tags": {
   "get": {
    "operationId": "driver_tags_api_v1_devices__device_id__mt76_driver_tags_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "window_s",
      "required": false,
      "schema": {
       "default": 3600,
       "maximum": 604800,
       "minimum": 5,
       "title": "Window S",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TagSummary"
         },
         "title": "Response Driver Tags Api V1 Devices  Device Id  Mt76 Driver Tags Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Driver Tags"
   }
  },
  "/api/v1/devices/{device_id}/mt76/instrumentation": {
   "get": {
    "operationId": "instrumentation_api_v1_devices__device_id__mt76_instrumentation_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "refresh",
      "required": false,
      "schema": {
       "default": false,
       "title": "Refresh",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Instrumentation"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Instrumentation"
   }
  },
  "/api/v1/devices/{device_id}/mt76/snapshot": {
   "get": {
    "operationId": "snapshot_api_v1_devices__device_id__mt76_snapshot_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "wake",
      "required": false,
      "schema": {
       "default": false,
       "title": "Wake",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Mt76Snapshot"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Snapshot"
   }
  },
  "/api/v1/devices/{device_id}/mt76/trace-aggregate": {
   "get": {
    "operationId": "trace_aggregate_api_v1_devices__device_id__mt76_trace_aggregate_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "window_s",
      "required": false,
      "schema": {
       "default": 300,
       "maximum": 86400,
       "minimum": 5,
       "title": "Window S",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TraceAggregate"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Trace Aggregate"
   }
  },
  "/api/v1/devices/{device_id}/pci-config/hexdump": {
   "get": {
    "operationId": "pci_hexdump_api_v1_devices__device_id__pci_config_hexdump_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Pci Hexdump Api V1 Devices  Device Id  Pci Config Hexdump Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Pci Hexdump"
   }
  },
...
  "/api/v1/patches/{patch_id}": {
   "get": {
    "operationId": "get_patch_api_v1_patches__patch_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "patch_id",
      "required": true,
      "schema": {
       "title": "Patch Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Patch"
   }
  },
...
 }
}